Refactored as requested
